fetched user data for profile
